moved btn and client side logic to mobile-menu
